Don't manually specify TypeMeta field
I originally started looking at this from the perspective of the test
examples, where something didn't look 100% to me, so I figured I'd
open a PR to start a discussion and figure out if my understanding is
correct or not.

Example:

Should the `Kind` in the following snippet not be `MemcachedList`
rather than `Memcached`?

```go
memcachedList := &operator.MemcachedList{
        TypeMeta: metav1.TypeMeta{
                Kind:       "Memcached",
                APIVersion: "cache.example.com/v1alpha1",
        },
}
```

I've seen the same thing in other operators, and I'm not sure if it's
the correct way, or if it's just wrong copied from this example.

As I think more about it, all of the correct information seems to be
inferrable from the type (`operator.MemcachedList`), so the `TypeMeta`
never needs to be added manually. Is that correct?

If it is, here's a pull request to remove it from everywhere that it's
manually specified here! :)